Move globals.js to state/ and update imports
