+++ index.html
@@ -404,7 +404,7 @@
         }
       }
     </script>
-    <script type="module" src="js/main.js"></script>
+    <script type="module" src="js/state/main.js"></script>
     <script type="module" src="js/modals/welcome.js"></script>
     <script type="module" src="js/modals/input.js"></script>
     <script type="module" src="js/modals/information.js"></script>
