+++ js/viz/draw/app.js
@@ -2,7 +2,7 @@ import { Application } from "pixi.js";
 import { Viewport } from "pixi-viewport";
 import { dragEnd } from "./drag.js";
 import { setRenderable } from "./renderable.js";
-import { getPixiState } from "../../globals.js";
+import { getPixiState } from "../../state/globals.js";
 import { objectColor } from "../types/vizStyles.js";
 
 const pixi = getPixiState();
